Добавил проекту модульность
