+++ index.html
@@ -233,7 +233,7 @@
   </template>
 
   <!-- Скрипт-->
-  <script src="./js/main.js"></script>
-  
+  <script src="./js/main.js" type="module"></script>
+
 </body>
 </html>
